requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
passlib[bcrypt]
python-dotenv
python-multipart
